Replace sample r2.data file with updated version
Based on sister changes to reddit-i18n, r2.data
should be JSON formatted. This is only used when
reddit-i18n translations are not in use (for opensource
clones/staging). In addition, the fallback
I18N_PATH (used when reddit-i18n repo not found)
was incorrect.

Also, remove unused imports in translation.py