Restructure sr bar clipping to avoid Android browser auto scaling.
The default Android browser of various flavors automatically scales the
page beyond the defined viewport width because of our abnormally wide
top bar. I juggled some of the CSS to be able to absolutely position the
bar within the page, which seems to avoid triggering this auto scaling
behavior.

The my subreddits dropdown needed some restyling to avoid clipping the
dropdown, when shown.