Add static domain middleware for testing compiled CSS.
In production, static files are located in the root of the static file
host. On a local reddit install, they're only available in /static/.
This middleware allows you to test the production static files layout in
a development environment.

If debug=true and uncompressedJS=false are set in the ini, this middleware
simulates a static files host by prepending /static/ to the URL before
handing the request off to the static files app.

To use this for testing purposes, add an additional domain alias like
redditstatic.local to your hosts file and set static_domain in the ini.
Compiled static files will then load from the app using this middleware.