Move stub admin-templates to root template directory.
This fixes the import of stub templates in open-source installs.

Before pluginization, the admin-stub templates were imported from the
admin sub-module of the templates directory if r2admin were not
available. Since we're not doing that logic anymore, these stub
templates need to move into the root template directory so they can be
found by the system.