Get rid of awkward and underused utils.to_js.
Buttonlite didn't really do anything with this function since it
disabled callback *and* escape.

The response wrapper fixes a bug where the callback is a unicode object
and since the content object is UTF-8 encoded things go south ("'ascii'
codec can't decode byte ...").

This also removes a few scattered imports of the function that were
unused.