cookies: Don't encrypt the unlogged-user _options cookie.
This cookie is part of the CDN cache key. Since the encrypted blob uses
a random salt, this essentially makes requests with the cookie
unchacheable. The cookie does not contain any sensitive data (just a
list of languages and a boolean True/False). Decrypting and normalizing
the order of params means that the cache key will be more useful. We
also get a fair reduction in size of the cookie.