Use consistent scaling on the timings graph
This lets one, at a glance, get a feel for how long a page took
to load. Timing data is scaled such that browser-width is 8 seconds
(unless the page took longer than 8 seconds). Clicking on the timings
will zoom the graph in for better investigation of what parts took
what amounts of time.