error_mapper: Don't redirect to ErrorController for pagecached errors.
This prevents a situation where we get a cached error (e.g. a 404) back
from the pagecache and then the error_mapper (called by the
ErrorDocuments middleware) sees that status code and redirects the app
to the ErrorController which then re-renders the error and inserts it
into the pagecache at the end of the request.