Fix multi details sidebar collection initial fetch and render.
The switch from collection reset -> set caused the change handler in
MultiDetails to immediately fire off a refreshListing.

This change causes the collection contents to be loaded before the
instantiation of the view (due to preload), and makes the view handle
initial rendering of the collection contents in its constructor.