Temporary bandaid for top bar breaking various subreddits.
sr_pops has put certain subreddits into the top subreddit list multiple
times. This causes there to be multiple buttons in the srbar. The sort
logic in NavMenu.find_selected short circuited when the lists were of
length 1, but now that they aren't it's dying rightfully on this
attribute not existing on SubredditButton. This is a temporary patch to
help until the sr_pops issue is fixed.