Line up rank and midcol elements using a <style> element.
This is a hack to scale the .rank and .midcol elements consistently
across the spotlight box and link listings. We don't know the maximum
link score length until we get to the link listing, at which point we
output a <style> tag to space things globally on the page. It's horribly
ugly, but maybe not as ugly as the style tags and jQuery measurement
hacks it replaces...