Overhaul scraper_q processing.
This is primarily about removing the dead old scrapers in favor of a
streamlined oEmbed-based system.  The new embed.ly scraper uses their
Service API to determine which URLs can be scraped through them. This
removes the giant list of domains and regexes in the code and makes the
scraper_q proc always have the latest list of scrapable domains.