Start writing HTTPS-friendly subreddit stylesheets.
This does several things to subreddit stylesheets:
- stores them on the thumbs buckets rather than the main static bucket.
  (this was not desirable before on reddit.com due to CDN configuration)
- enforces a new restriction of custom (%%style%%) images only in
  stylesheets to make secure urls easier to resolve. existing subreddits
  are grandfathered in for now.
- writes, if possible as above, a second stylesheet that references
  subreddit images over https.

At some point in the future, the thumbs buckets should be directly accessible
on HTTPS via the same URLs which would remove the need for the second
stylesheet to be created and uploaded. The custom image rules and other changes
would still be good.