Use `cythonize` to manage translation of .pyx -> .c.
This removes the need to run `make pyx` before running setup.py on an
initial install which had stopped working due to various imports needed
for the makefile. Additionally, cythonize is capable of tracking
dependencies the cython files have, such as external headers or other
cython files.

I had to upgrade setuptools to distribute to get v0.6.16 which is the
minimum required to properly use `cythonize` due to a bug in setuptools:

https://bitbucket.org/tarek/distribute/issue/195/