Make ListingChooser full-height on short pages.
The JS check in ListingChooser.initialize only takes care of the case
where the list of multis is longer than the body. However, when there
are very few multis *and* the page content is short, you get a weird
truncated listing chooser. This fixes that by setting a min-height on
body which works because html is already height: 100%.