Gold goal: switch to dynamic goal
Previously the goal was simply a static amount, this changes it so that
it is set to 110% of the average revenue from the previous 7 days
(disregarding highest and lowest revenues). To prevent too large of
swings due to "spikes", the goal is restricted to no more than a 20%
change from the previous day's goal.