Update recommend endpoint and validation.
get_recommendations was recently changed to accept id36s instead of subreddit
objects for the omit list. This change updates the recommender api endpoint
to match. It also makes the VSRByNames() validator return an empty dict in
case of error so it's safe to call .values() on the result.

Note: Fixes the broken "more suggestions" link in multi recommendations.