install-reddit: Focus on dev-environment installs and reorganize.
Production installs should work pretty differently and I have a few
plans for that.  In the meantime, this simplifies away a few of the bits
that were ill advisedly trying to help production installs and changes
the installer to be better suited to the modern environment with plugins
etc. by putting all code in a src/ directory.