Search box: fix right alignment
The containing div for the search box is 300 pixels wide.  The search box has 1
pixel of border and 2 pixels of padding.  Since the padding and border are on
both the left and right sides, we get 300 - 2 - 4 = 294px as the ideal width of
the box.  However, it *has* been 295, which causes a very slight alignment
issue with everything else in the sidebar.

Using box-sizing includes the border and padding in the width calculation, so
we don't have to do these sorts of calculations in the future, which is ideal.
Now things should look ever so slightly better. :)