Post submit: reinstate disallowed type feedback
We got [a bug report][0] that attempting to submit a link to a text-only
subreddit (or vice versa) was failing silently.  This turns out to have been an
unintended consequence of 5da814b.

Somewhat unintuitively, `JsonResponse`'s `has_errors()` method modifies state
in the response object; if one of the errors you tell it to check for is found
in the global `c.errors`, it caches that determination in an instance variable.
This variable is then what's looked at in `make_response()`, which is called by
the `validatedForm()` decorator we put on the endpoint.

So, when we return early from a controller without making the appropriate
`has_errors()` call, the error doesn't make its way to the end user, and that's
exactly what was happening.

This change addresses the immediate issue by ditching the `sr` check entirely
and instead relying on `VSubmitSR` (and checking for all the errors it can
add).  However, the larger problem of flow and unexpected state mutation is
still there - mostly because I'm not sure how to go about dealing with it.

[0]: http://www.reddit.com/r/bugs/comments/2jiwsb/there_is_no_user_feedback_when_you_attempt_to/