New windows pref: disable default-on for phones
See 0b7deff for some background.

This ends up being particularly annoying on phones, so we're changing the
default back for them.  An overall revert or an inclusion of tablets are both
stil in the cards, but for right now we need to fix this.

This affects both the desktop site and .compact.

@umbrae gets the credit for writing this; I just tested it and am committing.