Add proxying_for param for proxy gilding accounts
Allows proxy accounts to send the username of the user that the
gilding is being done on behalf of. This will still use a creddit
from the proxy account, but everything else related to the gilding will
be associated with the actual purchaser, to enable things like replying
to the gilding messages, update the gilder's trophy and /gilded page,
etc.