Improve moderator info page title
Since this page is accessible to non-mods, it doesn't make much sense to title
it "edit moderators" - people who can't edit anything can see it.

Also, while I was messing around in here, I noticed the titles we put on these
pages are awfully generic.  I didn't want to touch too many things, so we're
only adding a subreddit suffix to this page.