Report form: hide on gold payment page
In reddit/reddit@76858eb, I expanded the places we could use the report form.
Unfortunately, it got caught by a CSS rule on the reddit gold payment page that
made it visible, despite not being at all useful on that page. :)

Now that CSS rule is slightly more scoped.  The only other `<form>` that's
inside `.gold-payment` but **not** `.gold-wrap` is the share dialog, which has
an inline style hiding it.