Use react.js version with bundled addons.
This does _not_ bump the version number of react, it just uses the version
with addons bundled in.  These are mostly utility features, but the main
reason is for the animations addon.  Trying to add CSS transitions to elements
entering or exiting the DOM _without_ this addon would be a major pain.