Reports: fix in modmail
When a user reports a message, we've been checking to make sure it's actually
sent to them.  However, when a message is sent to a subreddit's moderation
team, any member of the mod team should be able to report the message, despite
it not being sent to *them*, directly.

We now extend the check to also check for moderator status if necessary.

[Reported][0] by /u/psdtwk.

[0]: https://www.reddit.com/r/bugs/comments/2p6m56/i/cmuw61w