Consistencize account text, part 2
While looking at reddit/reddit#1043, I was reminded that the work in
reddit/reddit@d2e2cef was incomplete, so I did a bit of `ack`ing and came up
with this set of changes.  This should be a pretty good set of changing login,
logged in, log out, logged out, etc. to the "sign out" equivalents, and
"register an account" and friends to "create an account".

There are a bunch of variable names and endpoints around that I didn't touch,
because that would break all sorts of things.  This *should* be only including
text directly shown to users (usually passed through `_()`) or indirectly shown
via the API docs.