markdown: Make headers inherit color.
If a subreddit wants to change the default text color (e.g. for a dark theme)
they'd need to override the color for the .md element and each header tag
(h1-h6) individually.  This makes it so the header tags inherit from the .md
element to make this less painful.