Robots: allow crawling of mobile pages
We got a message from Google Webmaster Tools letting us know they'd like to
pretend to be a mobile device and crawl our mobile pages, but can't, because we
disallow it in our `robots.txt` rules.  We have canonicalization set up to
point towards desktop pages, so we shouldn't be hurt by allowing this.

We should also at some point provide two-way annotations to let them go from
desktop -> mobile[0], but for now this should suffice.

[0]: https://developers.google.com/webmasters/mobile-sites/mobile-seo/configurations/separate-urls