Thumbnail scraper: handle protocol-relative urls
`urllib2` by default doesn't handle protocol-relative urls, nor can it without
access to the context in which the url was created.  This has been causing
problems with scraping thumbnails from pages where the `og:image` attribute
uses a protocol-relative url, as reported in reddit/reddit#1069 and
[/r/bugs](http://redd.it/2q62g2).

Now we coerce any protocol-relative urls we extract from the page into the same
protocol as the page itself.