vote: Stop dual-writing to postgres.
This is the final step in the saga of the pg vote rel destruction. We've
been dual-writing to PG and C* while gaining confidence in the pure-C*
model being able to survive full load.  This kills the pgvote databases
and moves forward in a pure-cassandra world for votes which should save
us considerable operational headaches.  After rolling this out, we can
not switch back without considerable effort.

When he reached the New World, Cortez burned his relational databases.
As a result his queue processors were well motivated.