selfserve: Fix sponsor domain override .
When the new promo form went from being a normal ajax call to a multipart form, it
stopped using the 'post_pseudo_form' method.  That method ignored inputs with the
'gray' class applied, which was being used for displaying the 'optional' placeholder
string.  This was probably created before there was widespread support for the
placeholder html attribute.