Add UrlParser.switch_subdomain_by_extension()
Previously, the only choices for changing a desktop url into a mobile one (or
vice versa, or going to XML, etc.) were to do it manually or to override
`c.render_style` before calling `add_sr()`.  Now there's an easier way through
`UrlParser`.

This necessitated moving the mappings from subdomains to hinted render
extensions out of the local namespace of a middleware function and into a more
accessible place, so now we also have access to it for other miscellaneous
uses.