Count skipped items in UserController.keep_fn.
This may give us information about how frequently incorrect data
is read from Cassandra LinkVotesByAccount and CommentVotesByAccount.
When a user votes on a link the link is put into a cached query and
a column is written to LinkVotesByAccount. LinkVotesByAccount is read
at consistency level ONE so it can return bad data. We have probably
received bad data if we were given a link by the liked cached query
but don't see a vote for it in LinkVotesByAccount.