Split uploaded media into separate buckets
We've operated for a while with just one set of s3 buckets for storing static
media (thumbnails, stylesheets, images uploaded for stylesheets).  However, due
to the additional-hop nature of image previews through imgix before reaching
s3, we're finding it'd be convenient to have imgix-served images in a separate
s3 bucket.

The media providers now take a *category* on `put()`.  We have a mapping from
category to bucket names, which will (at least for the time being) remain
`s3_media_buckets` for everything other than image previews.