Image previews: fix old urls
We've changed the url structure of image previews a number of times, which
breaks everything uploaded prior to the latest version.  This script should
find all preview images that have been uploaded thus far, move them to the
appropriate place, and save an updated and correct storage url in every Link
that uses them.