Add explicit sorting preference
For a long time, reddit has had a hidden comment sort preference - when you
changed your sort using the menu, it would stick and stay as the default.  This
has lead to a lot of confusion, as people will change their sort for one
thread, then not realize that's used for all the rest (and so wonder why, say,
all the comments they're viewing have pretty low scores, because they don't
realize they're sorting by `new` now instead of `best`).

Now we're getting rid of that behavior and replacing it with an explicit
preference.

From a user perspective, the ideal way to handle this would be for us to use
the same backend; that would keep any sorts that people have set.  However, all
that logic was genericized into `VMenu` to deal with all sorts of menus called
from all sorts of controllers, and that ended up making the key to query
something rather incidental (`front_sort`).  I'm also not a fan of validators
having side effects like that, so I just gave up detangling `VMenu` and used
the more common `VOneOf` and a preference that fits in with the rest of user
preferences.

Additionally, we suspect many users who have changed their sorts to something
other than best have done so accidentally (that is, intending for it to be
temporary), so starting afresh might actually be a better option, anyways.  If
not, we can write a backfill to make this migration.