AutoMod RuleTarget __init__: don't be destructive
Since part of RuleTarget.__init__() involves deleting items from the
values dict, this actually changed the dict that was passed to it, which
could have side effects back in the overall Rule creation. Specifically,
this was causing the approve_banned initialization to fail because it
tried to check for the presence of a key that had been destroyed by
this.