AutoMod: error if invalid search check for type
If all of the fields in a search check are invalid for a type, we should
throw an error instead of just quietly disregarding that search check.
The previous behavior caused a rule with, for example, separate searches
for both "domain" and "body" to apply to comments, even though the
presence of a domain check should have caused it to be restricted to
only submissions.