Open Graph: use preview images
We use the [Open Graph protocol][0] to hint to Facebook and other sites what
images they should show as an on-site preview of reddit content.  For most of
our comments pages, that ends up being the thumbnail image, which is pretty
small.

We're now collecting preview images that are sometimes larger than the
thumbnails, so let's start using them when we can!

[0]: http://ogp.me/