AutoMod: add support for aliases on RuleComponents
This also rearranges the init logic for RuleTargets slightly, which will
have a side-effect of strengthening the validation a little bit. For
example, after this change it will not be possible to set "is_top_level"
on a submission, since that only makes sense for comments. Previously it
would have validated but been silently ignored.