AutoMod: fix overly-greedy match placeholder regex
The regex to find {{match}}-type placeholders is currently being
overly-greedy in cases with multiple placeholders in the same string.
For example, something like "Your post was removed because {{match}} is
in the title, please remove {{match}} from the title and resubmit." will
fail due to the regex spanning from the start of the first placeholder
until the end of the second. Replacing the "." in the regex with "[^}]"
will ensure that the matches stay confined to the same placeholder.