CombinedSearch: Fix sort filter menus.
The js that controls the showing/hiding of the menus expect each dropdown
element to be the only instance amongst its siblings.  In other words, each
dropdown button and list must be wrapped in a container element for it to work
correctly.

Additionally, this moves the filter menus to the right side of the header for
better visibility.