AutoMod: fix RuleTarget matches not being cleared
A RuleTarget's matches dict (which stores the results of the search
checks on that item) was only being cleared in the __init__ function, so
it was actually being retained across checks against multiple different
items. This meant that {{match}} placeholders would occasionally be
replaced by the result of a match from an entirely different item. This
change makes it so that the matches will be cleared at the start of each
check instead.