Cleanup: Remove unused javascript.
A pass at cleaning up some frontend code by removing unused functions.  All of
these functions are either 1. not referenced anywhere or 2. referenced
somewhere, but don't actually _do_ anything. In the latter case, the reference
is removed as well.

The list of functions removed:

- hover_open_menu
- update_user
- emptyInput
- clear_all_langs
- check_some_langs
- show_unfriend
- pure_domain
- parse_domain
- save_usertext
- save_thing
- unsave_thing
- $.fn.same_author
- tool_cover
- fire_success
- check_name
- too_big

Additionally, the 'success_field' template in utils.html is removed, as it is
imported several places but never rendered (and was the only place the
fire_success function was being called).