TEMPORARY: Prevent expando buttons added by RES from triggering expando behavior.
http://redd.it/392zol

Expandos added by extensions can be mistaken for the ExpandoLink's own button, resulting in opening parallel expandos (reddit's and the add-on) or closing an expando by clicking a nested expando. This is avoided by verifying the clicked button is owned by the ExpandoLink.

Extensions can avoid this by stopping propagation on clicks to their expando buttons.