Renotify unspammed comments
If a comment is unspammed, put a notification back in the inbox
of all possible recipients (parent, username notifications, etc).

Keep track of the `new` status when the comment has been marked as spam.
When it is unspammed, try to get the previous `new` status so that it
isn't marked as unread if it's already been read.