Open graph: up max image size to 320px wide
With a little bit of testing, we decided 320px-wide Open Graph images are ideal
for Facebook.  Specifically, they seem to crop down to 160px squares, except
when presenting images to retina devices, in which case you'll see 320x320.
So, if we can produce a 320px-wide image, let's provide it, and if not, we'll
fall back to a smaller size before trying other sources.