Add a mod log entry for replaced stickies
Previously, when stickying a post, the mod log would only include an
entry for the stickying of the new post, even if it replaced an existing
sticky. In the case of a replacement, this adds a second log entry
indicating which post was unstickied to make room for the new one. In
the case of accidentally stickying something, this can make it much
easier to locate the previously-stickied post.