Create quarantined subreddits content gate
When a user visits a quarantined subreddit for the first time, they are
presented with a content gate so that they can choose to optin to view
the subreddit's content. This preference is remembered per subreddit.
If a user is logged out, it requires the user to log in before validating
that the user has chosen to opt in to view the subreddit.