Timeouts: add api info endpoint
This allows app developers to pull information about the user's ban from
`/user/<user>/about.json`.  This is useful not only for disabling the UI for
certain features, but also because we want to display prominent notices telling
the user they're in timeout, and we mostly have to rely on app devs to do that
for us outside of reddit.com.