Timeouts: Add modal for forbidden actions.
Works similarly to the 'login-required' modal.  Links that perform actions that are
forbidden when 'in timeout' are given the 'access-required' class, which triggers
the modal if the current user is in timeout.  Some links additionally needed to have
their default click handlers edited to exit early in this case (e.g. the ynbutton's toggle function).