setup.py: Remove check for distribute>=0.6.16.
This check was in place for c20db7e11d6bbcfc79830f96a4dee99f078cd4e0 to make
sure that we were using a new enough version of distribute to handle
the "cythonize" command. distribute was a now-deprecated fork of setuptools,
and the versions of setuptools on precise and trusty are new enough to not
need this check.