Remove the brick from the submit page
On the submit page, there is a brick before "please be mindful of
reddit's few rules and follow good reddiquette".

Since the rules page has been replaced with the content policy and
the section that explains the brick is no longer around, this reference
doesn't make sense any more.

This commit removes the brick, removes the title attribute from the
notice's div, and updates the link to go directly to reddit's new
content policy.