Track which subreddits have front-page-able posts
A user's front page will only include posts that are newer than
g.HOT_PAGE_AGE, so there's no point including subreddits in the front
page set if they haven't had a post inside that period. This commit
creates a new CF and adds a value to it (with a TTL equal to
g.HOT_PAGE_AGE) whenever a new post is made in a subreddit.

This should be deployed separately and left to collect data for at least
g.HOT_PAGE_AGE, then the following commit that actually uses this data
can be deployed.

Potential future enhancement: also update this data on post deletion,
removal, and approval since there are cases where the result from this
won't correctly match up with whether the subreddit actually has a valid
post for inclusion. This only tracks whether one was *made* during the
time period, not whether it's visible.