Exclude inactive subreddits from front page sets
When selecting which subreddits should be included in a user's front
page, this makes sure to only include subreddits if they have had a post
made recently enough to be included in the front page. This will prevent
wasting "front page slots" on subreddits that won't contribute anything
to the page anyway.