Fix error with other_theme through OAuth prefs API
The OAuth prefs API was removing any preference validators where the
attribute wasn't in Account._defaults. The only thing this affected was
pref_other_theme, which caused an other_theme setting to come through as
a string instead of a Subreddit since it was no longer going through the
VSrByName validator. This would cause a crash on the .can_view check in
filter_prefs, since that assumes that the theme subreddit is an actual
Subreddit object.