Thing: convert _date to g.tz timezone.
When loaded from postgres, thing._date.tzinfo will be a
psycopg2.tz.FixedOffsetTimezone object. Storing the thing in memcache
causes it to be pickled, and there may be a bug in the unpickling of
psycopg2.tz.FixedOffsetTimezone objects. Converting the date to
g.tz makes it a pytz.UTC object.