Ads: Treat subreddits included in nsfw collections as nsfw.
NSFW ads can be targeted at nsfw collections, however the
subreddits in those collections may or may not be properly marked
as nsfw.

This ensures any subreddit that is apart of a nsfw collection
will not used to target a user on a sfw listing even if they are
not marked nsfw; as well as the opposite.