Fix VoteDetailsByDay.count_votes()
This was still using the old method of calling _rowkey (which took a
date argument). The date needs to be converted to a datetime, and can
then be used to call _rowkey_by_datetime().

I also switched to use .get_count() instead of manually counting the
columns from the row, which should probably be faster.