Remove mobile settings for create subreddit
The mobile settings `key_color` and `related_subreddits` are not
included in the create subreddit form. Remove the fields from the new
subreddit in order to workaround a bug in subreddit creation when
`feature_mobile_settings` is enabled.

The `Subreddit.related_subreddits` depends on `Subreddit.name`
for constructing `Subreddit._related_multipath`, but depending on the
order that the attributes are set on the new object, `Subreddit.name`
may not actually be available yet.