Events: re-add truncation logic inside publisher
Even the new maximum event size limit may not be large enough due to
unicode characters. So this commit re-adds the ability to do event
truncation, but moves into inside the EventPublisher instead of the
events themselves. This makes more sense, since individual events
shouldn't need to (or even be able to) know the maximum size that their
publisher supports.