Uncollapse all comment threads with distinguished descendants.
In sticky comments, we hide comment threads which are children of
a sticky comment to avoid derailing a conversation. However, a
distinguished comment within that chain can override that (as it
should). This means we need to walk up the ancestor list and
de-collapse the thread that leads to this comment.

We were already doing this for Q&A sort which also has special
hiding rules, we just need it to now apply to all sorts.