add_comments: Set a timeout for acquiring the lock
The most common case for an exception in this block is when a
deploy occurs and causes the job server to restart all the services,
including the commentstree_q consumer. When this happens the consumer
is killed while holding onto the lock, so when a new consumer starts
up it tries to get the lock and fails. Adding a timeout will allow
the old lock to expire while the new consumer retries.