Event publisher: fixes to limiting event size
There were a couple of errors with the way the maximum event size was being
limited that this fixes:

* Only the maximum batch size was being accounted for, not the maximum
  size of an individual event (which is 20% of the batch limit).
* When truncation was done, we were adding an is_truncated field to the
  event, but the truncation didn't account for the size of this new
  field.

This commit fixes these issues and moves the truncation-handling into a
new wrapper class called PublishableEvent, instead of doing it inside
the queue-processor itself. It also takes advantage of the
application_headers support on amqp messages to send info about which
field is truncatable separate from the actual event data. This lets us
avoid needing to deserialize the JSON unless truncation is actually
necessary (and supported).

Graphite events are also added so that we can more easily track how
often oversize events need to be truncated or dropped.