install-reddit: Don't sudo inside install-reddit, require it outside
Previously, install-reddit.sh ran install/reddit.sh with sudo. This
confused install.cfg when install-reddit.sh was itself run with sudo
because the inner sudo would cause $SUDO_USER to erroneously be root.
install-reddit.sh _should_ be run as sudo so that when it prints out
install.cfg the values are accurate, but should not run
install/reddit.sh with an additional sudo since that'll mess things up.