new_comment: rearrange mutator creation
This moves the creation of a CachedQueryMutator inside
update_comment_notifications, to simplify the process of calling it
directly.

Related to that, the CachedQueryMutator inside new_comment is
now only created when it's needed in the (rare) spam-related cases,
since it no longer needs to create one to be able to call
update_comment_notifications.