Expose feature flags data via API
We need to access experiment variant assignments in the mobile web
client, and in general it would be useful to condition client features
on server-provided feature flags. We expose only the enabled flags and
only the experiments for which non-null variants have been assigned. We
expose the feature data through /api/me.json and /api/v1/me.