Link.affects_karma_type: remove is_promoted_post
This check is related to the code for the upcoming "User Promoted Posts"
(see https://redd.it/4upf11), which hasn't been open-sourced yet. So
this attr doesn't exist yet, and isn't working. This should be removed
for now but can be added back in when the rest of the code is
open-sourced.