PermalinkCommentOrderer: handle ancestor missing from tree
PermalinkCommentOrderer will walk up the comment tree if context is
requested (i.e. context=2 means show the two ancestors of the requested
comment). If an ancestor comment is missing from the tree we will now
fail cleanly and fall back to the regular CommentOrderer.

This situation can arise when there are multiple commentstree_q consumers,
causing a child comment to be added to the comment tree before its parent.