AutoMod: fix bug with int account threshold checks
It should be valid to define a check for one of the account thresholds
similar to:

    author:
        comment_karma: 0

to check for exactly that value. However, since this is parsed in as an
int, trying to do a regex match on it (to check for the presence of an
operator - equal/greater-than/less-than sign) crashes.

This ensures that the comparison value is converted to a string so that
the rest of the code can handle it.