utils.timetext: use long to handle larger periods
timetext() deals with seconds, and it is currently giving weird results
for any time period that's longer than the max int value of 2147483647
seconds (about 68 years). The total number of seconds was already being
declared as a long, but one of the intermediate variables used in the
calculations was still an int and caused this issue.

This switches that variable to long as well, which resolves the problem.
The unused variable `count2` was also removed.